Removes debugging left in CallBack.
This removes code I left in to print out in LogCat if the application
failed to retrieve the commit data from the server. This does not use
Crouton to show an error message if it failed to connect, because the
diff CallBack already does this, and if it failed to connect to the
server, it would display two error messages.