Adds menu option to disable text wrap
This feature adds a checkable menu option to toggle wrapping lines of
code and commit messages to the width of the screen in DiffActivity.

This was achieved by creating a custom HorizontalScrollView with an
onMeasure method that measures its child with an exact MeasureSpec of
its own width. If wrapping is disabled, the super method is called,
making it act just like a normal HorizontalScrollView.